Unconditionally serialize FFTW plan creation
FFTW plans must always be created serially (a known restriction of the
library). Yandasoft has a mutex for this, but it's only declared/used
when compiling the code with OpenMP support, and thus other types of
multi-threaded applications eventually fail.

Yandasoft was originally designed to use OpenMP to provide parallelism
on tasks, but now that it's distributed in a more reusable fashion this
assumption might not hold anymore. One example of this is the
integration of yandasoft with DALiuGE, where multiple tasks might run
concurrently within the same process, but in different threads.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>