Indexing derivates by integer
One of the major costs of GenericNormalEquations::add is the continuous
lookup of the derivative values living within the ComplexDiff objects,
which in turn are the cells of the input ComplexDiffMatrix. These
lookups are first of all string-based, resulting in many calls to a
map.find method where keys are std::strings, and thus hurting
performance. Many of these lookups are also duplicated, and thus
performance worsens even more.

This patch introduces an integer-based indexing approach to alleviate
this situation. Both sets of derivative values (real and imaginary) are
first collected into a plain vector that is indexed by parameter number,
row and column, and thus can be quickly accessed later using these
coordinates. This results in a small memory cost increase, but in a big
reduction on computation, as the string-based lookups occur only once
when filling up the vectors.