Removing duplicate map lookups
Several methods of the Param class perform an initial lookup on the
underlying map to check if the name being queried is present in the map
or not. However, retrieving the actual value undergoes a second lookup,
leading to inefficiencies. This patch removes *some* (not all) of these
unnecessary duplicate lookups, improving the runtime of the system.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>