Inlining addModel*Product fully-indexed functions
The main user of these functions will be the PreAvgCalBuffer class in
yandasoft, who currently creates a slice out of a PolXProducts object
and then adds products into the 1D slice; we are planning to move that
logic to avoid slices and add products directly into the parent object,
and that will give us a higher boost if these routines can be inlined,
making the compiler be more efficient on the code that gets generated.

The addModel*Product function overloads that take less parameters are
those currently used by the "slices" objects, and apparently are not
used anywhere else, so it doesn't make sense to inline these as well as
they will become deprecated/unused rather soon.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>