Using non-parallel, MPI-enabled SparseMatrix
Following the newly set convention on SparseMatrix.getComm(), if
MPI_COMM_NULL is returned it means that the matrix object is not being
used in a parallel fashion, even though it was compiled with MPI
support.

Most of the code changed in this commit previously assumed that
getComm() returned a valid communicator, and acted accordingly. With
this patch we are changing those checks to act differently: instead of
failing upon finding an MPI_COMM_NULL we now behave the same way the
code behaves when no MPI support is found (i.e., serially).