Supporting non-parallel, MPI-enabled SparseMatrix
The SparseMatrix class has always had non-parallel support but only when
compiled without MPI; on the other hand, when compiled *with* MPI
support it always requires to be used in a parallel fashion.

This commit is the first in a series that seeks to support non-parallel
sparse matrices even when the software was built with MPI support.
SparseMatrix not only holds an MPI communicator, but also exposes it to
allow users share the same MPI context as the matrix. The convention is
that if matrix.getComm() returns MPI_COMM_NULL then the matrix should be
assumed to be working in a non-parallel fashion, and hence users should
act accordingly.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>