Using SparseMatrix in parallel only when required
Until now the creation of SparseMatrix objects in the LinearSolver class
always required MPI to be initialized, even when using MPI_COMM_SELF as
the communicator. With the latest changes introduced in the last commits
we can now confidently create a SparseMatrix object with no MPI
communicator, even if the software was built with MPI support. In these
cases, SparseMatrix stores MPI_COMM_NULL as its communicator, signaling
that it doesn't work in parallel.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>