Adding missing copy assignment operator
In LinearSolver we copy-assign SparseMatrix objects, but the underlying
MPI_Comm object is not simply copy-assignable and needs special care
instead.

This commit implements a copy-assignment operator for the SparseMatrix
class that knows how to deal with an MPI_Comm object; in the long term
however it might be better to replace this approach with a new type that
encapsulates only an MPI_Comm and knows how to move/copy it, so users
don't have to deal with these details.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>